Create file for EditTask function
